% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XcmsExperiment.R, R/XcmsExperimentHdf5.R
\name{XcmsExperimentStash}
\alias{XcmsExperimentStash}
\alias{saveMsObject,XcmsExperiment,PlainTextParam-method}
\alias{readMsObject,XcmsExperiment,PlainTextParam-method}
\alias{saveObject,XcmsExperiment-method}
\alias{saveMsObject,XcmsExperiment,AlabasterParam-method}
\alias{readMsObject,XcmsExperiment,AlabasterParam-method}
\alias{saveMsObject,XcmsExperimentHdf5,PlainTextParam-method}
\alias{readMsObject,XcmsExperimentHdf5,PlainTextParam-method}
\alias{saveObject,XcmsExperimentHdf5-method}
\alias{saveMsObject,XcmsExperimentHdf5,AlabasterParam-method}
\alias{readMsObject,XcmsExperimentHdf5,AlabasterParam-method}
\title{\code{XcmsExperiment} Stash}
\usage{
\S4method{saveMsObject}{XcmsExperiment,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperiment,PlainTextParam}(object, param, ...)

\S4method{saveObject}{XcmsExperiment}(x, path, ...)

\S4method{saveMsObject}{XcmsExperiment,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperiment,AlabasterParam}(object, param, ...)

\S4method{saveMsObject}{XcmsExperimentHdf5,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperimentHdf5,PlainTextParam}(object, param, ...)

\S4method{saveObject}{XcmsExperimentHdf5}(x, path, ...)

\S4method{saveMsObject}{XcmsExperimentHdf5,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperimentHdf5,AlabasterParam}(object, param, ...)
}
\arguments{
\item{object}{An \code{XcmsExperiment} or \code{XcmsExperimentHdf5} object.}

\item{param}{The parameter object to select and configure the stash format.
Either \link[MsStash:AlabasterParam]{MsStash::AlabasterParam} or \link[MsStash:PlainTextParam]{MsStash::PlainTextParam}.}

\item{...}{For \code{saveMsObject()}: optional arguments passed down to the
\code{saveMsObject()} function to stash the \code{Spectra} object (if present),
such as \code{consolidate}. For \code{readMsObject()}: optional arguments for the
\code{readMsObject()} call to restore the \code{Spectra} object (such as
\code{spectraPath}). See \link[SpectraStash:SpectraStash]{SpectraStash::SpectraStash} for more information.}

\item{x}{An \code{XcmsExperiment} or \code{XcmsExperimentHdf5} object.}

\item{path}{For \code{saveObject()}:}
}
\description{
\link[xcms:XcmsExperiment]{xcms::XcmsExperiment} or \link[xcms:XcmsExperimentHdf5]{xcms::XcmsExperimentHdf5} \emph{xcms} result objects
can be stored to (or read from) \emph{XcmsExperimentStash}es using the
\code{saveMsObject()} and \code{readMsObject()} functions which take a second argument
\code{parameter} to select and configure the format of the stash. The xcmsStash
extends the \link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} defined in the
\emph{MsExperimentStash} package, i.e., it shares the same content, but adds the
\emph{xcms} preprocessing results to it. A self-contained, portable, stash can
be created by passing \code{consolidate = TRUE} to the \code{saveMsObject()}
(or \code{saveObject()}) call. This will ensure that also the data files for the
full MS data are stored within the stash folder (resulting in an eventually
much larger stash folder size).

The supported stash formats are listed in the sections below.
}
\note{
Overwriting an existing \emph{XcmsExperimentStash} is not allowed.
}
\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


This alabster stash format is the most complete and reliable way for
long-term (and portable) storage of an \code{XcmsExperiment} or
\code{XcmsExperimentHdf5} object.
Objects can be saved or read from this stash format either using the
\code{saveMsObject()} and \code{readMsObject()} functions together with an
\code{AlabasterParam} object or also using the \code{\link[alabaster.base:saveObject]{alabaster.base::saveObject()}} and
\code{\link[alabaster.base:readObject]{alabaster.base::readObject()}} functions from the \emph{alabaster.base} package.
The alabaster stash format for \code{XcmsExperiment} and \code{XcmsExperimentHdf5}
objects extends the \link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash}.
Storage of the object's MS data and sample information is handled by the
functions of the \link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} with the \emph{xcms}
preprocessing results being added to the stash.
Data from the object's slots are stored to respective folders (using
alabaster functionality). Refer to the documentation of the
\link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} for information on the format of the
stored MS data and sample information. \emph{xcms}-specific data/folders are:

For \code{XcmsExperiment} objects:
\itemize{
\item \emph{chrom_peaks}: the identified chromatographic peaks (\code{chromPeaks()}),
stored as a numeric HDF5 array.
\item \emph{chrom_peak_data}: the data from the object's \code{chromPeakData()}, saved in
HDF5 format.
\item \emph{feature_definitions}: the object's \code{featureDefinitions()}, saved in HDF5
format.
\item \emph{xcms_experiment_process_history.json}: the object's \code{processHistory()}
serialized in JSON format.
}

For \code{XcmsExperimentHdf5} objects:
\itemize{
\item The HDF5 file with the preprocessing results is copied into the stash
folder (same file name as the original HDF5 file).
\item \emph{hdf5_file} folder: contains the \code{character(1)} with the file name of the
HDF5 file with the \emph{xcms} preprocessing results.
\item \emph{hdf5_mod_count} folder: contains the content of the object's
\verb{@hdf5_mod_count} slot (an \code{integer(1)}).
\item \emph{sample_id} folder: contains the \code{character} vector with the IDs of the
individual samples in the result object.
\item \emph{chrom_peaks_ms_level} folder: contains the \code{integer} vector with the MS
level(s) in which chromatographic peak detection(s) was/were performed.
\item \emph{gap_peaks_ms_level} folder: contains the \code{integer} vector with the MS
level(s) in which gap-filling was performed.
\item \emph{features_ms_level} folder: contains the \code{integer} vector with the MS
levels for which features were defined.
\item \emph{xcms_experiment_process_history.json} file: the object's
\code{processHistory()} serialized in JSON format.
}
}

\section{Text file-based format, \code{PlainTextParam}}{


\code{saveMsObject()}/\code{readMsObject()} with \code{PlainTextParam} allows to save and
read \emph{xcms} result objects from a text file-based stash. All preprocessing
results are stored in separate tabulator delimited text files. Each of these
files has the row names (representing either the chromatographic peak IDs or
feature IDs) written as first column and the column names as first row. The
files grouped by preprocessing result are:

For \code{XcmsExperiment} objects:
\itemize{
\item Chromatographic peak detection results are stored in
\emph{xcms_experiment_chrom_peaks.txt} (\code{chromPeaks()}) and
\emph{xcms_experiment_chrom_peak_data.txt} (\code{chromPeakData()}).
\item Correspondence results: the definitions of the LC-MS features
(\code{featureDefinitions()}) are stored to
\emph{xcms_experiment_feature_definitions.txt} (one row per feature) and the
assignment between chromatographic peaks and features (the \verb{$peakidx}
column of \code{featureDefinitions()}) to
\emph{xcms_experiment_feature_peak_index.txt}, with one row for each mapping
of a chromatographic peak to a feature. These files are not created (or
present in the stash) if no correspondence analysis was performed.
\item The processing history (\code{processHistory()}) is stored in JSON format to
\emph{xcms_experiment_process_history.json}.
}

For \code{XcmsExperimentHdf5} objects: contents of the individual slots of the
object are saved using \code{writeLines()} to one text file per slot.
\itemize{
\item The HDF5 file with the preprocessing results is copied into the stash
folder (same file name as the original HDF5 file).
\item \emph{xcms_experiment_hdf5_hdf5_file.txt} file: the name of the HDF5 file with
the preprocessing results (content of the \verb{@hdf5_file} slot of the object).
\item \emph{xcms_experiment_hdf5_hdf5_mod_count.txt} file: the \code{integer(1)} from the
\verb{@hdf5_mod_count} slot of the object.
\item \emph{xcms_experiment_hdf5_sample_id.txt} file: the IDs of the individual
samples, saved one per line in the txt file.
\item \emph{xcms_experiment_hdf5_chrom_peaks_ms_level.txt} file: the MS level(s) in
which chromatographic peak detection was performed (one per line).
\item \emph{xcms_experiment_hdf5_gap_peaks_ms_level.txt} file: the MS level(s) in
which gap-filling was performed (one per line).
\item \emph{xcms_experiment_hdf5_featrures_ms_level.txt} file: the MS level(s) in
which features were defined (one per line).
\item \emph{xcms_experiment_process_history.json} file: the object's
\code{processHistory()} serialized in JSON format.
}
}

\examples{

## Load the example xcms result object
library(xcms)
library(Spectra)
library(MsExperiment)
xmse <- loadXcmsData()

## Define the path where to create the XcmsExperimentStash
d <- file.path(tempdir(), "xcms_stash")

## Save the XcmsExperiment to a stash in alabaster format; Note: with
## `consolidate = TRUE` also the MS data files are **copied** into the
## stash folder, creating a bigger, but portable and self-contained
## archive
saveMsObject(xmse, AlabasterParam(d), consolidate = TRUE)

## Show the content of the stash folder
library(fs)
dir_tree(d)

## Read the MsExperiment from the stash (without xcms preprocessing results)
res <- readMsObject(MsExperiment(), AlabasterParam(d))
res

## Read the full xcms result object
res <- readMsObject(XcmsExperiment(), AlabasterParam(d))

## Show the first identified chromatographic peaks
chromPeaks(res) |> head()
}
\author{
Philippine Louail, Johannes Rainer
}
