% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XcmsExperiment.R
\name{XcmsExperimentStash}
\alias{XcmsExperimentStash}
\alias{saveMsObject,XcmsExperiment,PlainTextParam-method}
\alias{readMsObject,XcmsExperiment,PlainTextParam-method}
\alias{saveObject,XcmsExperiment-method}
\alias{saveMsObject,XcmsExperiment,AlabasterParam-method}
\alias{readMsObject,XcmsExperiment,AlabasterParam-method}
\title{\code{XcmsExperiment} Stash}
\usage{
\S4method{saveMsObject}{XcmsExperiment,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperiment,PlainTextParam}(object, param, ...)

\S4method{saveObject}{XcmsExperiment}(x, path, ...)

\S4method{saveMsObject}{XcmsExperiment,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{XcmsExperiment,AlabasterParam}(object, param, ...)
}
\arguments{
\item{object}{A \code{XcmsExperiment} object.}

\item{param}{The parameter object to select and configure the stash format.
Either \link[MsStash:AlabasterParam]{MsStash::AlabasterParam} or \link[MsStash:PlainTextParam]{MsStash::PlainTextParam}.}

\item{...}{For \code{saveMsObject()}: optional arguments passed down to the
\code{saveMsObject()} function to stash the \code{Spectra} object (if present),
such as \code{consolidate}. For \code{readMsObject()}: optional arguments for the
\code{readMsObject()} call to restore the \code{Spectra} object (such as
\code{spectraPath}). See \link[SpectraStash:SpectraStash]{SpectraStash::SpectraStash} for more information.}

\item{x}{A \code{XcmsExperiment} object.}

\item{path}{For \code{saveObject()}:}
}
\description{
\code{XcmsExperiment} \emph{xcms} result objects can be stored to (or read from)
\emph{XcmsExperimentStash}es using the \code{saveMsObject()} and \code{readMsObject()}
functions which take a second argument \code{parameter} to select and configure
the format of the stash. The XcmsExperimentStash extends the
\link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} defined in the \emph{MsExperimentStash}
package, i.e., it shares the same content, but adds the \emph{xcms} preprocessing
results to it.

The supported stash formats are listed in the sections below.
}
\note{
Overwriting an existing \emph{XcmsExperimentStash} is not allowed.
}
\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


This alabster stash format is the most complete and reliable way for
long-term (and portable) storage of an \code{XcmsExperiment}. Objects can be
saved or read from this stash format either using the \code{saveMsObject()}
and \code{readMsObject()} functions together with an \code{AlabasterParam} object
or also using the \code{\link[alabaster.base:saveObject]{alabaster.base::saveObject()}} and
\code{\link[alabaster.base:readObject]{alabaster.base::readObject()}} functions from the \emph{alabaster.base} package.
The alabaster stash format for \code{XcmsExperiment} extends the
\link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash}. Storage of the object's MS data and
sample information is handled by the functions of the
\link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} with the \emph{xcms} preprocessing results
being added to the stash.
Data from the object's slots are stored to respective folders (using
alabaster functionality). Refer to the documentation of the
\link[MsExperimentStash:MsExperimentStash]{MsExperimentStash::MsExperimentStash} for information on the format of the
stored MS data and sample information. \emph{xcms}-specific data/folders are:
\itemize{
\item \emph{chrom_peaks}: the identified chromatographic peaks (\code{chromPeaks()}),
stored as a numeric HDF5 array.
\item \emph{chrom_peak_data}: the data from the object's \code{chromPeakData()}, saved in
HDF5 format.
\item \emph{feature_definitions}: the object's \code{featureDefinitions()}, saved in HDF5
format.
\item \emph{xcms_experiment_process_history.json}: the object's \code{processHistory()}
serialized as a JSON object.
}
}

\section{Text file-based format, \code{PlainTextParam}}{


\code{saveMsObject()}/\code{readMsObject()} with \code{PlainTextParam} allows to save and
read \emph{xcms} result objects from a text file-based stash. All preprocessing
results are stored in separate tabulator delimited text files. Each of these
files has the row names (representing either the chromatographic peak IDs or
feature IDs) written as first column and the column names as first row. The
files grouped by preprocessing result are:
\itemize{
\item Chromatographic peak detection results are stored in
\emph{xcms_experiment_chrom_peaks.txt} (\code{chromPeaks()}) and
\emph{xcms_experiment_chrom_peak_data.txt} (\code{chromPeakData()}).
\item Correspondence results: the definitions of the LC-MS features
(\code{featureDefinitions()}) are stored to
\emph{xcms_experiment_feature_definitions.txt} (one row per feature) and the
assignment between chromatographic peaks and features (the \verb{$peakidx}
column of \code{featureDefinitions()}) to
\emph{xcms_experiment_feature_peak_index.txt}, with one row for each mapping
of a chromatographic peak to a feature. These files are not created (or
present in the stash) if no correspondence analysis was performed.
\item The processing history (\code{processHistory()}) is stored in JSON format to
\emph{xcms_experiment_process_history.json}.
}
}

\examples{

## Load the example xcms result object
library(xcms)
library(Spectra)
library(MsExperiment)
xmse <- loadXcmsData()

## Ensure all parameter objects in the object's process history are in
## the latest format.
xmse@processHistory <- lapply(xmse@processHistory, function(z) {
    z@param <- updateObject(z@param)
    z
})

## Define the path where to create the XcmsExperimentStash
d <- file.path(tempdir(), "xcms_stash")

## Save the XcmsExperiment to a stash in alabaster format; Note: with
## `consolidate = TRUE` the MS data files are also **copied** into the
## stash
saveMsObject(xmse, AlabasterParam(d), consolidate = TRUE)

## Show the content of the stash folder
library(fs)
dir_tree(d)

## Read the MsExperiment from the stash (without xcms preprocessing results)
res <- readMsObject(MsExperiment(), AlabasterParam(d))
res

## Read the full xcms result object
res <- readMsObject(XcmsExperiment(), AlabasterParam(d))

## Show the first identified chromatographic peaks
chromPeaks(res) |> head()
}
\author{
Philippine Louail, Johannes Rainer
}
